tests/wpt/tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-nested.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/whatwg/html/issues/11880">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<select>
  <button></button>
  <option>one</option>
  <option>two</option>
</select>

<script>
test(() => {
  const select = document.querySelector('select');
  const button = document.querySelector('button');

  const parentSelectedcontent = document.createElement('selectedcontent');
  parentSelectedcontent.id = 'parent';
  parentSelectedcontent.textContent = 'parent';
  const childSelectedcontent = document.createElement('selectedcontent');
  childSelectedcontent.id = 'child';
  childSelectedcontent.textContent = 'child';

  // Appending the disconnected childSelectedcontent to the disconnected
  // parentSelectedcontent does not trigger any DOM changes.
  parentSelectedcontent.appendChild(childSelectedcontent);
  // Appending the pair of nested selectedcontent elements triggers cloning
  // for each of them. The child one is not updated because its post-insertion
  // steps look for an ancestor select element, but the post-insertion steps of
  // the parent one already removed it from the select element.
  button.appendChild(parentSelectedcontent);

  assert_equals(parentSelectedcontent.innerHTML, 'one',
    'parent selectedcontent should not get cleared after inserting.');
  assert_equals(childSelectedcontent.innerHTML, 'child',
    'child selectedcontent should not be modified after inserting.');
  assert_equals(childSelectedcontent.parentNode, null,
    'child selectedcontent should get disconnected after inserting.');

  // Appending childSelectedcontent to the connected parentSelectedcontent does
  // not trigger any DOM changes because childSelectedcontent is marked as
  // disabled for having an ancestor selectedcontent element.
  parentSelectedcontent.appendChild(childSelectedcontent);
  // Assigning to select.value triggers an update for descendant selectedcontent
  // elements. The child one is not updated because we check whether selectedcontent
  // elements are disabled when building the list of selectedcontent elements to update.
  select.value = 'two';

  assert_equals(parentSelectedcontent.innerHTML, 'two',
    'parent selectedcontent should get updated when selected option is changed.');
  assert_equals(childSelectedcontent.innerHTML, 'child',
    'child selectedcontent should not get updated when selected option is changed.');
  assert_equals(childSelectedcontent.parentNode, null,
    'child selectedcontent should get disconnected when selected option is changed.');
}, 'Modification behavior when nesting selectedcontent elements.');
</script>
