LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-nested-expected.txt



FAIL Modification behavior when nesting selectedcontent elements. assert_equals: parent selectedcontent should get cleared. expected 0 but got 1
